requirements.txt
Django==1.7
Unidecode==0.04.16
django-autoslug==1.7.2
pyjade==3.0.0
simplejson==3.6.3
six==1.8.0
sorl-thumbnail==11.12
wsgiref==0.1.2
